test(logging,storage): convert env-var-gated integration tests to build tags (Issue #1244)
Three test files used runtime env-var checks (CFGMS_TEST_TIMESCALEDB, CFGMS_TEST_SYSLOG,
DATABASE_URL) to gate tests that require live infrastructure. This pattern causes silent
t.Skip noise in plain `go test` runs and obscures real test results.

Changes:
- pkg/logging/providers/timescale/plugin_test.go: add //go:build integration tag;
  remove isTimescaleTestEnabled() helper and all 6 env-var t.Skip guards. The file
  is now entirely excluded from non-integration builds instead of silently skipping.
- pkg/logging/subscribers/syslog/subscriber_integration_test.go: new file with
  //go:build !windows && integration; contains TestSyslogSubscriber_HandleLogEntry
  (moved from subscriber_test.go) without the CFGMS_TEST_SYSLOG env-var check.
- pkg/logging/subscribers/syslog/subscriber_test.go: remove TestSyslogSubscriber_HandleLogEntry
  (now in integration file); all other unit tests remain unchanged.
- pkg/storage/providers/database/plugin_integration_test.go: new file with
  //go:build integration; contains TestDatabaseProvider_Integration (moved from
  plugin_test.go) without the DATABASE_URL env-var check; testing.Short() guard preserved.
- pkg/storage/providers/database/plugin_test.go: remove TestDatabaseProvider_Integration
  and unused pkg/storage/interfaces import.

After this change, `go test` on all three packages produces zero skips and zero
failures for the formerly-gated tests — they simply do not compile into the binary.
Running with -tags=integration compiles and lists them as expected.

Fixes #1244
